✨ Add comprehensive doc comment extraction across all 25 language extractors
Implements systematic documentation comment parsing for all supported languages,
improving semantic search accuracy and developer experience. Completed via 24
parallel TDD agents with full test coverage.

Key Features:
- Doc comment extraction for 25 languages (Bash, C, C++, C#, CSS, Dart, GDScript,
  Go, HTML, Java, JavaScript, Kotlin, Lua, PHP, PowerShell, Python, Razor, Regex,
  Ruby, Rust, SQL, Swift, TypeScript, Vue, Zig)
- Language-specific comment patterns (JSDoc, Doxygen, Python docstrings, etc.)
- Ancestor sibling scanning for complex AST structures (SQL, Dart)
- Comprehensive test suite (8 new test modules with 24+ test cases)

Code Quality:
- All 1139 tests passing (0 failures)
- Clippy warnings reduced from 66 → 20 (70% improvement)
- Fixed struct update syntax, borrow checker issues, case comparisons
- TDD methodology followed throughout (RED → GREEN → REFACTOR)

Architecture:
- BaseExtractor.find_doc_comment() used consistently across all extractors
- SymbolOptions.doc_comment field populated for all symbol types
- Multi-language comment pattern detection (block, line, language-specific)
- Non-breaking changes - backward compatible with existing symbols

Stats: 179 files changed, 5997 insertions(+), 964 deletions(-)

+++ src/tools/ast_symbol_finder.rs
@@ -175,7 +175,7 @@ impl ASTSymbolFinder {
         let kind = node.kind();
         let mut in_range = true;
         if let Some((start, end)) = byte_range {
-            in_range = node.start_byte() as usize >= start && node.end_byte() as usize <= end;
+            in_range = node.start_byte() >= start && node.end_byte() <= end;
         }
 
         if in_range
@@ -186,8 +186,8 @@ impl ASTSymbolFinder {
                 .unwrap_or(false)
         {
             occurrences.push(SymbolOccurrence {
-                start_byte: node.start_byte() as usize,
-                end_byte: node.end_byte() as usize,
+                start_byte: node.start_byte(),
+                end_byte: node.end_byte(),
                 node_kind: kind.to_string(),
                 context: SymbolContext::Comment,
                 line: node.start_position().row + 1,
